requirements.txt
alabaster==1.0.0
annotated-types==0.7.0
anyio==4.6.0
babel==2.16.0
bcrypt==4.2.0
black==24.8.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
com2ann==0.3.0
dnspython==2.6.1
docutils==0.21.2
email_validator==2.2.0
fastapi==0.115.0
fastapi-cli==0.0.5
h11==0.14.0
httpcore==1.0.6
httptools==0.6.1
httpx==0.27.2
idna==3.8
imagesize==1.4.1
Jinja2==3.1.4
libcst==1.4.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mypy-extensions==1.0.0
numpydoc==1.8.0
packaging==24.1
passlib==1.7.4
pathspec==0.12.1
platformdirs==4.3.3
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
PyJWT==2.9.0
python-dotenv==1.0.1
python-multipart==0.0.12
pyupgrade==3.17.0
PyYAML==6.0.2
requests==2.32.3
rich==13.9.2
ruff==0.6.5
shed==2024.3.1
shellingham==1.5.4
sniffio==1.3.1
snowballstemmer==2.2.0
Sphinx==8.0.2
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
starlette==0.38.6
tabulate==0.9.0
terminusdb-client==10.2.4
tokenize-rt==6.0.0
tqdm==4.66.5
typeguard==2.13.3
typer==0.12.5
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.31.0
uvloop==0.20.0
watchfiles==0.24.0
websockets==13.1

# Custom packages
-e ./elegant-notes-database
